# backend/app/static/swagger.yaml
openapi: 3.0.0
info:
  title: Employee Management API
  description: API to manage users, employees, contracts, etc.
  version: 1.0.0
servers:
  - url: http://localhost:5002
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
  /login:
    post:
      summary: Logs a user into the system
      requestBody:
        description: User credentials
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    type: object
                    properties:
                      username:
                        type: string
        '401':
          description: Invalid username or password

  /signup:
    post:
      summary: Registers a new user
      requestBody:
        description: User details
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                username:
                  type: string
                password1:
                  type: string
                password2:
                  type: string
      responses:
        '200':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '400':
          description: Invalid request or passwords mismatch

  /create_employee:
    post:
      summary: Creates a new employee
      security:
        - bearerAuth: []
      requestBody:
        description: Employee details
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                employee_name:
                  type: string
                company_name:
                  type: string
                start_date:
                  type: string
                job_title:
                  type: string
                salary:
                  type: number
      responses:
        '201':
          description: Employee created successfully
        '400':
          description: Invalid input
        '401':
          description: Unauthorized

  /employees_wo_contract:
    get:
      summary: Get employees without contracts
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    employee_name:
                      type: string
        '404':
          description: No employees found